Fix linking issues from prev commit
