+++ VoodooSMBus/Info.plist
@@ -62,16 +62,14 @@
 	<string>3.0</string>
 	<key>OSBundleLibraries</key>
 	<dict>
-		<key>com.apple.iokit.IOHIDFamily</key>
-		<string>2.0</string>
 		<key>com.apple.iokit.IOPCIFamily</key>
 		<string>2.9</string>
 		<key>com.apple.kpi.iokit</key>
-		<string>18.5</string>
+		<string>15</string>
 		<key>com.apple.kpi.libkern</key>
-		<string>18.5</string>
+		<string>15</string>
 		<key>com.apple.kpi.mach</key>
-		<string>18.5</string>
+		<string>15</string>
 	</dict>
 	<key>OSBundleRequired</key>
 	<string>Root</string>
